requirements.txt
pyspark>=3.4.0
pyyaml>=6.0
pandas>=2.0.0
lxml>=4.9.0
pytest>=7.0.0
